+++ release_notes.md
@@ -1,1 +1,6 @@
 - updated urls and imports
+- Removed acquirer code and use the bascule acquirer instead
+- Removed webhook registerer code and use the wrp-listener registerer instead
+- Use bascule to authenticate the hash with middleware instead of in the primaryHandler
+- Updated yaml and documentation
+- Changed webhook request config url to just be the server portion of the url